HSEARCH-3429 Add a dedicated test for ORM full-text search queries
